+++ src/app/manifestGeneration/manifestGenerators/generatorDevPreview/TabManifestUpdater.ts
@@ -4,6 +4,7 @@
 
 import { IManifestUpdater } from "../../IManifestUpdater";
 import { GeneratorTeamsAppOptions } from "../../../GeneratorTeamsAppOptions";
+import { v4 as uuid } from 'uuid';
 
 export class TabManifestUpdater implements IManifestUpdater {
 
@@ -11,7 +12,7 @@ export class TabManifestUpdater implements IManifestUpdater {
     public updateManifest(manifest: any, options: GeneratorTeamsAppOptions): void {
         if (options.tabType == "static") {
             (<any[]>manifest.staticTabs).push({
-                entityId: "default-data",
+                entityId: uuid(),
                 name: options.tabTitle,
                 contentUrl: `https://{{HOSTNAME}}/${options.tabName}/?name={loginHint}&tenant={tid}&group={groupId}&theme={theme}`,
                 scopes: ["personal"]